Rainwater Harvesting and Groundwater Recharge Household Survey (BASEflow, 2025)

This dataset contains household-level survey data collected in 2025 under the BASEflow monitoring and data systems using the mWater digital data collection platform. The survey was conducted across selected communities to assess household water access conditions, perceptions of groundwater systems, and adoption of rainwater harvesting (RWH) practices.

Usage

rainwaterhsurvey

Format

A tibble with 479 rows and 18 variables

district

Name of district where the household is located

community

Name of the local community or village of the respondent

primary_water_source

Main source of water used by the household for drinking and domestic purposes

water_scarcity

Whether the household experiences water scarcity

scarcity_months

Months of the year when the household experiences water scarcity

perceived_water_source

Respondents perception of where their household water originates from

belief_underground_rivers

Whether the respondent believes that underground rivers supply boreholes and wells

belief_groundwater_exhausted

Whether the respondent believes that drying of boreholes indicates complete depletion of groundwater

belief_groundwater_replenished

Whether the respondent believes that groundwater can be replenished or naturally refilled

belief_recharge_detectable

Respondents view on whether it is possible to detect or know if groundwater is being replenished

rwh_practice

Whether the household practices rainwater harvesting

rwh_barrier

Main reason why the household does not practice rainwater harvesting

rwh_use

Primary uses of harvested rainwater within the household

belief_rwh_recharge

Respondents belief that rainwater harvesting contributes to groundwater recharge

belief_rwh_expensive

Respondents belief that rainwater harvesting requires expensive technology or is unaffordable

belief_rwh_unreliable

Respondents belief that rainwater harvesting is ineffective due to rainfall variability or unreliability

recharge_solutions

Respondents suggestions on methods or practices that can help replenish groundwater

rwh_support_needed

Types of support, information, or interventions that would encourage adoption of rainwater harvesting or groundwater recharge practices